# service/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  AttentionItem:
    properties:
      date:
        type: string
      file:
        $ref: '#/definitions/File'
      id:
        type: integer
      time:
        type: string
      user:
        $ref: '#/definitions/AttentionUser'
    type: object
  AttentionListResponse:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/AttentionItem'
        type: array
    type: object
  AttentionUser:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  CreateDocRequest:
    properties:
      children_position_index:
        description: 子节点的位置
        type: integer
      content:
        type: string
      doc_name:
        type: string
      father_id:
        description: 父节点 id
        type: integer
      title:
        type: string
    type: object
...
  DeleteDocRequest:
    properties:
      doc_name:
        type: string
      project_name:
        type: string
    type: object
  DeleteFileRequest:
    properties:
      file_name:
        type: string
      project_name:
        type: string
    type: object
...
  FeedItem:
    properties:
      action:
        type: string
      date:
        type: string
      id:
        type: integer
      show_divider:
        description: 分割线
        type: boolean
      source:
        $ref: '#/definitions/Source'
      time:
        type: string
      user:
        $ref: '#/definitions/FeedUser'
    type: object
  FeedListResponse:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/FeedItem'
        type: array
    type: object
  FeedUser:
    properties:
      avatar_url:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  File:
    properties:
      file_creator:
        $ref: '#/definitions/AttentionUser'
      file_id:
        type: integer
      file_kind:
        type: integer
      file_name:
        type: string
      file_project_id:
        type: integer
      file_project_name:
        type: string
    type: object
...
  Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  Source:
    properties:
      id:
        type: integer
      kind:
        description: 类型，1 -> 团队，2 -> 项目，3 -> 文档，4 -> 文件，6 -> 进度（5 不使用）
        type: integer
      name:
        type: string
      project_id:
        type: integer
      project_name:
        type: string
    type: object
...
  UpdateDocRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
...
  UpdateFileRequest:
    properties:
      name:
        type: string
      url:
        type: string
    type: object
  UpdateFolderRequest:
    properties:
      name:
        type: string
    type: object
...
host: work.test.muxi-tech.xyz
info:
  contact: {}
  description: The gateway of muxi-workbench
  title: muxi-workbench-gateway
  version: "1.0"
paths:
...
  /attention/list/{id}:
    get:
      consumes:
      - application/json
      description: 获取此用户的关注list
      parameters:
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: last_id
        in: query
        name: last_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AttentionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: get attention list api
      tags:
      - attention
...
  /feed/list:
    get:
      consumes:
      - application/json
      description: 获取此用户的动态list
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: last_id
        in: query
        name: last_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/FeedListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: get feed list api
      tags:
      - feed
...
  /file/doc:
    post:
      consumes:
      - application/json
      description: 通过父节点 id，和插入节点位置确定新建文档的位置。
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: create_doc_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/CreateDocRequest'
      - description: project_id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: create a new doc api
      tags:
      - project
  /file/doc/{id}:
    delete:
      consumes:
      - application/json
      description: 删除doc
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: delete_doc_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/DeleteDocRequest'
      - description: doc_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: delete a doc api
      tags:
      - project
    put:
      consumes:
      - application/json
      description: 修改文档
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: doc_id
        in: path
        name: id
        required: true
        type: integer
      - description: update_doc_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/UpdateDocRequest'
      - description: project_id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: update doc info api
      tags:
      - project
...
  /file/file/{id}:
    delete:
      consumes:
      - application/json
      description: 删除文件
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: delete_file_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/DeleteFileRequest'
      - description: file_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: delete a file api
      tags:
      - project
    put:
      consumes:
      - application/json
      description: 修改文件
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 此文件 id
        in: path
        name: id
        required: true
        type: integer
      - description: update_file_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/UpdateFileRequest'
      - description: 此文件所属项目 id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: update file api
      tags:
      - project
...
  /folder/docfolder/{id}:
    delete:
      consumes:
      - application/json
      description: 删除文档文件夹
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: doc_folder_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: delete a doc folder api
      tags:
      - project
    put:
      consumes:
      - application/json
      description: 修改文档夹名字
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文档夹 id
        in: path
        name: id
        required: true
        type: integer
      - description: update_folder_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/UpdateFolderRequest'
      - description: 此文档夹所属项目 id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: update doc folder's name api
      tags:
      - project
...
swagger: "2.0"
...